Fix unfailable test.

+++ test/ondemand/unit_tests/test_fast_cache_assessment.sh
@@ -3,4 +3,8 @@ rm -rf ../test_data/fast_cache_testing/
 cp -r ../test_data/fast_cache ../test_data/fast_cache_testing
 
 python unit_tests/fast_cache_assessment.py
+if [[ "$1" != "0" ]];
+then
+    exit 1
+fi
 rm -rf ../test_data/fast_cache_testing/
